Additional Cogs, Enums and Classes
Added CommandDisability Enum for specifying in which way the command is
filtered.
Added CommandDisabledException to be used in checks, so they can be
easily filtered off.
Added GuildConfigCog that would handle the connection to mongoDB and
retrieving and storing data for GuildConfig.
Added MotorchikBot class that would attach a method to the commands.Bot
class that is a shorthand for getting a guild config using guild config
cog.